requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
black==22.12.0
Brotli==1.0.9
click==8.1.3
dark==0.7.1
decorator==5.1.1
dj-database-url==1.2.0
Django==4.1.5
django-cpf==0.1.0
django-crontab==0.7.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
executing==1.2.0
gunicorn==20.1.0
ipdb==0.13.11
ipython==8.8.0
jedi==0.18.2
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.10.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.6.2
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tomli==2.0.1
traitlets==5.8.0
wcwidth==0.2.5
whitenoise==6.3.0
